refactor(cli): collapse lifecycle verbs to positional (start/stop/kill/restart) (step 5)
