+++ cmd/kuke/kill/kill.go
@@ -17,48 +17,101 @@
 package kill
 
 import (
+	"fmt"
 	"strings"
 
-	cellcmd "github.com/eminwux/kukeon/cmd/kuke/kill/cell"
+	"github.com/eminwux/kukeon/cmd/config"
+	kukeshared "github.com/eminwux/kukeon/cmd/kuke/shared"
+	"github.com/eminwux/kukeon/internal/errdefs"
+	"github.com/eminwux/kukeon/pkg/api/kukeonv1"
+	v1beta1 "github.com/eminwux/kukeon/pkg/api/model/v1beta1"
 	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
 )
 
-// NewKillCmd builds the `kuke kill` parent command and registers all resource
-// kill subcommands. Persistent flags defined on the root kuke command are
-// inherited automatically via Cobra's command tree.
+// MockControllerKey is used to inject a mock kukeonv1.Client via context in tests.
+type MockControllerKey struct{}
+
+// NewKillCmd builds the `kuke kill <name>` leaf command. Cell is the only
+// resource this verb targets, so the noun is implied by the verb.
 func NewKillCmd() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:     "kill [name]",
-		Aliases: []string{"k"},
-		Short:   "Kill Kukeon resources (cell)",
-		Run: func(cmd *cobra.Command, _ []string) {
-			_ = cmd.Help()
+		Use:           "kill <name>",
+		Aliases:       []string{"k"},
+		Short:         "Immediately force-kill a cell",
+		Args:          cobra.ExactArgs(1),
+		SilenceUsage:  true,
+		SilenceErrors: false,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			name := strings.TrimSpace(args[0])
+			realm := strings.TrimSpace(viper.GetString(config.KUKE_KILL_CELL_REALM.ViperKey))
+			space := strings.TrimSpace(viper.GetString(config.KUKE_KILL_CELL_SPACE.ViperKey))
+			stack := strings.TrimSpace(viper.GetString(config.KUKE_KILL_CELL_STACK.ViperKey))
+
+			if realm == "" {
+				return fmt.Errorf("%w (--realm)", errdefs.ErrRealmNameRequired)
+			}
+			if space == "" {
+				return fmt.Errorf("%w (--space)", errdefs.ErrSpaceNameRequired)
+			}
+			if stack == "" {
+				return fmt.Errorf("%w (--stack)", errdefs.ErrStackNameRequired)
+			}
+
+			doc := v1beta1.CellDoc{
+				APIVersion: v1beta1.APIVersionV1Beta1,
+				Kind:       v1beta1.KindCell,
+				Metadata:   v1beta1.CellMetadata{Name: name, Labels: map[string]string{}},
+				Spec: v1beta1.CellSpec{
+					ID:      name,
+					RealmID: realm,
+					SpaceID: space,
+					StackID: stack,
+				},
+			}
+
+			client, err := resolveClient(cmd)
+			if err != nil {
+				return err
+			}
+			defer func() { _ = client.Close() }()
+
+			result, err := client.KillCell(cmd.Context(), doc)
+			if err != nil {
+				return err
+			}
+
+			cellName := result.Cell.Metadata.Name
+			if cellName == "" {
+				cellName = name
+			}
+			stackName := result.Cell.Spec.StackID
+			if stackName == "" {
+				stackName = stack
+			}
+			cmd.Printf("Killed cell %q from stack %q\n", cellName, stackName)
+			return nil
 		},
 	}
 
-	cmd.ValidArgsFunction = completeKillSubcommands
+	cmd.Flags().String("realm", "", "Realm that owns the cell")
+	_ = viper.BindPFlag(config.KUKE_KILL_CELL_REALM.ViperKey, cmd.Flags().Lookup("realm"))
+	cmd.Flags().String("space", "", "Space that owns the cell")
+	_ = viper.BindPFlag(config.KUKE_KILL_CELL_SPACE.ViperKey, cmd.Flags().Lookup("space"))
+	cmd.Flags().String("stack", "", "Stack that owns the cell")
+	_ = viper.BindPFlag(config.KUKE_KILL_CELL_STACK.ViperKey, cmd.Flags().Lookup("stack"))
 
-	cmd.AddCommand(
-		cellcmd.NewCellCmd(),
-	)
+	cmd.ValidArgsFunction = config.CompleteCellNames
+	_ = cmd.RegisterFlagCompletionFunc("realm", config.CompleteRealmNames)
+	_ = cmd.RegisterFlagCompletionFunc("space", config.CompleteSpaceNames)
+	_ = cmd.RegisterFlagCompletionFunc("stack", config.CompleteStackNames)
 
 	return cmd
 }
 
-// completeKillSubcommands provides shell completion for kill subcommand names.
-func completeKillSubcommands(_ *cobra.Command, _ []string, toComplete string) ([]string, cobra.ShellCompDirective) {
-	subcommands := []string{"cell"}
-
-	if toComplete == "" {
-		return subcommands, cobra.ShellCompDirectiveNoFileComp
+func resolveClient(cmd *cobra.Command) (kukeonv1.Client, error) {
+	if mockClient, ok := cmd.Context().Value(MockControllerKey{}).(kukeonv1.Client); ok {
+		return mockClient, nil
 	}
-
-	matches := make([]string, 0, len(subcommands))
-	for _, subcmd := range subcommands {
-		if strings.HasPrefix(subcmd, toComplete) {
-			matches = append(matches, subcmd)
-		}
-	}
-
-	return matches, cobra.ShellCompDirectiveNoFileComp
+	return kukeshared.DaemonClientFromCmd(cmd)
 }
