+++ cmd/kuke/start/start.go
@@ -17,48 +17,101 @@
 package start
 
 import (
+	"fmt"
 	"strings"
 
-	cellcmd "github.com/eminwux/kukeon/cmd/kuke/start/cell"
+	"github.com/eminwux/kukeon/cmd/config"
+	kukeshared "github.com/eminwux/kukeon/cmd/kuke/shared"
+	"github.com/eminwux/kukeon/internal/errdefs"
+	"github.com/eminwux/kukeon/pkg/api/kukeonv1"
+	v1beta1 "github.com/eminwux/kukeon/pkg/api/model/v1beta1"
 	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
 )
 
-// NewStartCmd builds the `kuke start` parent command and registers all resource
-// start subcommands. Persistent flags defined on the root kuke command are
-// inherited automatically via Cobra's command tree.
+// MockControllerKey is used to inject a mock kukeonv1.Client via context in tests.
+type MockControllerKey struct{}
+
+// NewStartCmd builds the `kuke start <name>` leaf command. Cell is the only
+// resource this verb targets, so the noun is implied by the verb.
 func NewStartCmd() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:     "start [name]",
-		Aliases: []string{"sta"},
-		Short:   "Start Kukeon resources (cell)",
-		Run: func(cmd *cobra.Command, _ []string) {
-			_ = cmd.Help()
+		Use:           "start <name>",
+		Aliases:       []string{"sta"},
+		Short:         "Start a cell",
+		Args:          cobra.ExactArgs(1),
+		SilenceUsage:  true,
+		SilenceErrors: false,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			name := strings.TrimSpace(args[0])
+			realm := strings.TrimSpace(viper.GetString(config.KUKE_START_CELL_REALM.ViperKey))
+			space := strings.TrimSpace(viper.GetString(config.KUKE_START_CELL_SPACE.ViperKey))
+			stack := strings.TrimSpace(viper.GetString(config.KUKE_START_CELL_STACK.ViperKey))
+
+			if realm == "" {
+				return fmt.Errorf("%w (--realm)", errdefs.ErrRealmNameRequired)
+			}
+			if space == "" {
+				return fmt.Errorf("%w (--space)", errdefs.ErrSpaceNameRequired)
+			}
+			if stack == "" {
+				return fmt.Errorf("%w (--stack)", errdefs.ErrStackNameRequired)
+			}
+
+			doc := v1beta1.CellDoc{
+				APIVersion: v1beta1.APIVersionV1Beta1,
+				Kind:       v1beta1.KindCell,
+				Metadata:   v1beta1.CellMetadata{Name: name, Labels: map[string]string{}},
+				Spec: v1beta1.CellSpec{
+					ID:      name,
+					RealmID: realm,
+					SpaceID: space,
+					StackID: stack,
+				},
+			}
+
+			client, err := resolveClient(cmd)
+			if err != nil {
+				return err
+			}
+			defer func() { _ = client.Close() }()
+
+			result, err := client.StartCell(cmd.Context(), doc)
+			if err != nil {
+				return err
+			}
+
+			cellName := result.Cell.Metadata.Name
+			if cellName == "" {
+				cellName = name
+			}
+			stackName := result.Cell.Spec.StackID
+			if stackName == "" {
+				stackName = stack
+			}
+			cmd.Printf("Started cell %q from stack %q\n", cellName, stackName)
+			return nil
 		},
 	}
 
-	cmd.ValidArgsFunction = completeStartSubcommands
+	cmd.Flags().String("realm", "", "Realm that owns the cell")
+	_ = viper.BindPFlag(config.KUKE_START_CELL_REALM.ViperKey, cmd.Flags().Lookup("realm"))
+	cmd.Flags().String("space", "", "Space that owns the cell")
+	_ = viper.BindPFlag(config.KUKE_START_CELL_SPACE.ViperKey, cmd.Flags().Lookup("space"))
+	cmd.Flags().String("stack", "", "Stack that owns the cell")
+	_ = viper.BindPFlag(config.KUKE_START_CELL_STACK.ViperKey, cmd.Flags().Lookup("stack"))
 
-	cmd.AddCommand(
-		cellcmd.NewCellCmd(),
-	)
+	cmd.ValidArgsFunction = config.CompleteCellNames
+	_ = cmd.RegisterFlagCompletionFunc("realm", config.CompleteRealmNames)
+	_ = cmd.RegisterFlagCompletionFunc("space", config.CompleteSpaceNames)
+	_ = cmd.RegisterFlagCompletionFunc("stack", config.CompleteStackNames)
 
 	return cmd
 }
 
-// completeStartSubcommands provides shell completion for start subcommand names.
-func completeStartSubcommands(_ *cobra.Command, _ []string, toComplete string) ([]string, cobra.ShellCompDirective) {
-	subcommands := []string{"cell"}
-
-	if toComplete == "" {
-		return subcommands, cobra.ShellCompDirectiveNoFileComp
+func resolveClient(cmd *cobra.Command) (kukeonv1.Client, error) {
+	if mockClient, ok := cmd.Context().Value(MockControllerKey{}).(kukeonv1.Client); ok {
+		return mockClient, nil
 	}
-
-	matches := make([]string, 0, len(subcommands))
-	for _, subcmd := range subcommands {
-		if strings.HasPrefix(subcmd, toComplete) {
-			matches = append(matches, subcmd)
-		}
-	}
-
-	return matches, cobra.ShellCompDirectiveNoFileComp
+	return kukeshared.ClientFromCmd(cmd)
 }
